# cuda : improve multi-GPU performance using cuBLAS