WIP in progress

Some results on 2x RTX A6000 without NVLink:

```bash
nvidia-smi topo -m
	GPU0	GPU1	CPU Affinity	NUMA Affinity	GPU NUMA ID
GPU0	X 	NODE	0-31,64-95	0		N/A
GPU1	NODE	X 	0-31,64-95	0		N/A

Legend:

  X    = Self
  SYS  = Connection traversing PCIe as well as the SMP interconnect between NUMA nodes (e.g., QPI/UPI)
  NODE = Connection traversing PCIe as well as the interconnect between PCIe Host Bridges within a NUMA node
  PHB  = Connection traversing PCIe as well as a PCIe Host Bridge (typically the CPU)
  PXB  = Connection traversing multiple PCIe bridges (without traversing the PCIe Host Bridge)
  PIX  = Connection traversing at most a single PCIe bridge
  NV#  = Connection traversing a bonded set of # NVLinks


LLAMA_CUBLAS=1 make -j batched-bench && ./batched-bench ./models/codellama-13b/ggml-model-q4_k.gguf 4096 1 99 1 512,3200 128,800 1

### pre-#3776

CUDA_VISIBLE_DEVICES=0,   Q4_K

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.266 |  1927.23 |    1.372 |    93.29 |    1.638 |   390.78 |
|   512 |    800 |    1 |   1312 |    0.256 |  1998.92 |    9.339 |    85.67 |    9.595 |   136.74 |
|  3200 |    128 |    1 |   3328 |    1.940 |  1649.80 |    2.324 |    55.07 |    4.264 |   780.50 |
|  3200 |    800 |    1 |   4000 |    1.939 |  1649.95 |   15.270 |    52.39 |   17.210 |   232.43 |

CUDA_VISIBLE_DEVICES=0,1, Q4_K

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.372 |  1374.82 |    1.953 |    65.56 |    2.325 |   275.28 |
|   512 |    800 |    1 |   1312 |    0.334 |  1532.62 |   12.881 |    62.11 |   13.215 |    99.28 |
|  3200 |    128 |    1 |   3328 |    2.402 |  1332.46 |    2.899 |    44.15 |    5.301 |   627.80 |
|  3200 |    800 |    1 |   4000 |    2.312 |  1384.32 |   18.865 |    42.41 |   21.177 |   188.89 |

CUDA_VISIBLE_DEVICES=0,   F16

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.145 |  3541.27 |    3.035 |    42.18 |    3.179 |   201.31 |
|   512 |    800 |    1 |   1312 |    0.143 |  3587.72 |   19.615 |    40.78 |   19.758 |    66.40 |
|  3200 |    128 |    1 |   3328 |    1.175 |  2723.75 |    3.980 |    32.16 |    5.155 |   645.60 |
|  3200 |    800 |    1 |   4000 |    1.177 |  2719.28 |   25.665 |    31.17 |   26.842 |   149.02 |

CUDA_VISIBLE_DEVICES=0,1  F16

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.471 |  1087.39 |    2.829 |    45.25 |    3.300 |   193.96 |
|   512 |    800 |    1 |   1312 |    0.461 |  1111.19 |   18.185 |    43.99 |   18.645 |    70.37 |
|  3200 |    128 |    1 |   3328 |    3.125 |  1024.04 |    3.761 |    34.03 |    6.886 |   483.31 |
|  3200 |    800 |    1 |   4000 |    3.185 |  1004.83 |   24.282 |    32.95 |   27.467 |   145.63 |


### #3776

CUDA_VISIBLE_DEVICES=0,   Q4_K

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.169 |  3034.15 |    1.304 |    98.13 |    1.473 |   434.43 |
|   512 |    800 |    1 |   1312 |    0.157 |  3259.24 |    8.404 |    95.20 |    8.561 |   153.26 |
|  3200 |    128 |    1 |   3328 |    1.355 |  2362.34 |    1.663 |    76.95 |    3.018 |  1102.72 |
|  3200 |    800 |    1 |   4000 |    1.356 |  2360.72 |   10.687 |    74.86 |   12.042 |   332.17 |

CUDA_VISIBLE_DEVICES=0,1  Q4_K

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.484 |  1058.74 |    1.891 |    67.71 |    2.374 |   269.57 |
|   512 |    800 |    1 |   1312 |    0.452 |  1132.24 |   11.824 |    67.66 |   12.276 |   106.88 |
|  3200 |    128 |    1 |   3328 |    3.016 |  1060.96 |    2.213 |    57.84 |    5.229 |   636.44 |
|  3200 |    800 |    1 |   4000 |    3.039 |  1052.82 |   14.169 |    56.46 |   17.208 |   232.44 |


CUDA_VISIBLE_DEVICES=0,   F16

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.146 |  3518.15 |    2.946 |    43.45 |    3.091 |   207.04 |
|   512 |    800 |    1 |   1312 |    0.139 |  3674.25 |   18.648 |    42.90 |   18.787 |    69.84 |
|  3200 |    128 |    1 |   3328 |    1.178 |  2717.46 |    3.308 |    38.69 |    4.486 |   741.88 |
|  3200 |    800 |    1 |   4000 |    1.177 |  2719.18 |   20.961 |    38.17 |   22.138 |   180.69 |

CUDA_VISIBLE_DEVICES=0,1  F16

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.503 |  1018.78 |    2.761 |    46.37 |    3.263 |   196.13 |
|   512 |    800 |    1 |   1312 |    0.481 |  1063.36 |   17.191 |    46.54 |   17.673 |    74.24 |
|  3200 |    128 |    1 |   3328 |    3.171 |  1009.03 |    3.069 |    41.71 |    6.240 |   533.34 |
|  3200 |    800 |    1 |   4000 |    3.136 |  1020.36 |   19.453 |    41.13 |   22.589 |   177.08 |


### PR

CUDA_VISIBLE_DEVICES=0,   Q4_K

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.169 |  3036.74 |    1.306 |    98.00 |    1.475 |   433.99 |
|   512 |    800 |    1 |   1312 |    0.159 |  3225.44 |    8.400 |    95.24 |    8.559 |   153.29 |
|  3200 |    128 |    1 |   3328 |    1.357 |  2358.35 |    1.663 |    76.97 |    3.020 |  1102.03 |
|  3200 |    800 |    1 |   4000 |    1.354 |  2363.78 |   10.683 |    74.89 |   12.037 |   332.32 |

CUDA_VISIBLE_DEVICES=0,1  Q4_K

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.420 |  1218.12 |    1.863 |    68.70 |    2.284 |   280.26 |
|   512 |    800 |    1 |   1312 |    0.456 |  1122.10 |   11.740 |    68.14 |   12.196 |   107.57 |
|  3200 |    128 |    1 |   3328 |    2.786 |  1148.80 |    2.195 |    58.32 |    4.980 |   668.22 |
|  3200 |    800 |    1 |   4000 |    2.755 |  1161.39 |   14.031 |    57.02 |   16.787 |   238.28 |

CUDA_VISIBLE_DEVICES=0,   F16

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.143 |  3591.50 |    2.941 |    43.53 |    3.083 |   207.58 |
|   512 |    800 |    1 |   1312 |    0.141 |  3639.13 |   18.628 |    42.95 |   18.769 |    69.90 |
|  3200 |    128 |    1 |   3328 |    1.179 |  2714.52 |    3.316 |    38.60 |    4.495 |   740.33 |
|  3200 |    800 |    1 |   4000 |    1.175 |  2722.29 |   20.974 |    38.14 |   22.149 |   180.59 |

CUDA_VISIBLE_DEVICES=0,1  F16

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.405 |  1265.49 |    2.746 |    46.62 |    3.150 |   203.16 |
|   512 |    800 |    1 |   1312 |    0.412 |  1244.13 |   17.146 |    46.66 |   17.557 |    74.73 |
|  3200 |    128 |    1 |   3328 |    2.793 |  1145.63 |    3.071 |    41.68 |    5.864 |   567.49 |
|  3200 |    800 |    1 |   4000 |    2.974 |  1075.91 |   19.411 |    41.21 |   22.385 |   178.69 |



LLAMA_CUBLAS=1 make -j && CUDA_VISIBLE_DEVICES=0   ./batched-bench models/codellama-7b/ggml-model-q4_k.gguf 18432 0 99 1 100 128 1,2,3,4,5,6,7,8,16,32,64
LLAMA_CUBLAS=1 make -j && CUDA_VISIBLE_DEVICES=0,1 ./batched-bench models/codellama-7b/ggml-model-q4_k.gguf 18432 0 99 1 100 128 1,2,3,4,5,6,7,8,16,32,64

### #3776

main: n_kv_max = 18432, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.082 |  1214.68 |    1.247 |   102.62 |    1.330 |   171.48 |
|   100 |    128 |    2 |    456 |    0.086 |  2314.49 |    1.876 |   136.44 |    1.963 |   232.33 |
|   100 |    128 |    3 |    684 |    0.119 |  2528.85 |    1.929 |   199.06 |    2.048 |   334.04 |
|   100 |    128 |    4 |    912 |    0.148 |  2708.72 |    1.985 |   257.89 |    2.133 |   427.56 |
|   100 |    128 |    5 |   1140 |    0.174 |  2878.71 |    3.137 |   204.00 |    3.311 |   344.31 |
|   100 |    128 |    6 |   1368 |    0.233 |  2575.44 |    3.191 |   240.68 |    3.424 |   399.54 |
|   100 |    128 |    7 |   1596 |    0.244 |  2869.75 |    3.271 |   273.91 |    3.515 |   454.05 |
|   100 |    128 |    8 |   1824 |    0.282 |  2841.72 |    3.336 |   306.98 |    3.617 |   504.25 |
|   100 |    128 |   16 |   3648 |    0.599 |  2671.80 |    6.453 |   317.37 |    7.052 |   517.31 |
|   100 |    128 |   32 |   7296 |    1.366 |  2342.92 |   11.914 |   343.81 |   13.279 |   549.42 |
|   100 |    128 |   64 |  14592 |    3.710 |  1725.04 |   17.319 |   473.01 |   21.029 |   693.90 |

main: n_kv_max = 18432, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.140 |   715.55 |    1.769 |    72.34 |    1.909 |   119.42 |
|   100 |    128 |    2 |    456 |    0.250 |   798.48 |    2.483 |   103.09 |    2.734 |   166.80 |
|   100 |    128 |    3 |    684 |    0.347 |   864.55 |    2.567 |   149.62 |    2.914 |   234.76 |
|   100 |    128 |    4 |    912 |    0.387 |  1033.48 |    2.620 |   195.41 |    3.007 |   303.27 |
|   100 |    128 |    5 |   1140 |    0.425 |  1175.38 |    3.036 |   210.79 |    3.462 |   329.33 |
|   100 |    128 |    6 |   1368 |    0.563 |  1065.55 |    3.044 |   252.31 |    3.607 |   379.27 |
|   100 |    128 |    7 |   1596 |    0.695 |  1006.70 |    3.202 |   279.82 |    3.897 |   409.50 |
|   100 |    128 |    8 |   1824 |    0.739 |  1082.45 |    3.277 |   312.50 |    4.016 |   454.20 |
|   100 |    128 |   16 |   3648 |    1.402 |  1141.31 |    5.404 |   378.95 |    6.806 |   535.98 |
|   100 |    128 |   32 |   7296 |    2.865 |  1117.02 |    9.097 |   450.26 |   11.962 |   609.95 |
|   100 |    128 |   64 |  14592 |    7.006 |   913.53 |   20.187 |   405.80 |   27.193 |   536.60 |

### PR

main: n_kv_max = 18432, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.082 |  1224.11 |    1.251 |   102.35 |    1.332 |   171.12 |
|   100 |    128 |    2 |    456 |    0.082 |  2439.71 |    1.876 |   136.45 |    1.958 |   232.87 |
|   100 |    128 |    3 |    684 |    0.116 |  2589.06 |    1.927 |   199.26 |    2.043 |   334.80 |
|   100 |    128 |    4 |    912 |    0.147 |  2728.94 |    1.977 |   259.01 |    2.123 |   429.51 |
|   100 |    128 |    5 |   1140 |    0.171 |  2929.65 |    3.121 |   205.07 |    3.292 |   346.35 |
|   100 |    128 |    6 |   1368 |    0.229 |  2623.49 |    3.177 |   241.74 |    3.406 |   401.68 |
|   100 |    128 |    7 |   1596 |    0.246 |  2845.25 |    3.265 |   274.39 |    3.511 |   454.52 |
|   100 |    128 |    8 |   1824 |    0.279 |  2872.48 |    3.345 |   306.11 |    3.624 |   503.35 |
|   100 |    128 |   16 |   3648 |    0.595 |  2689.10 |    6.425 |   318.75 |    7.020 |   519.64 |
|   100 |    128 |   32 |   7296 |    1.367 |  2341.62 |   11.887 |   344.59 |   13.253 |   550.51 |
|   100 |    128 |   64 |  14592 |    3.695 |  1732.04 |   17.321 |   472.95 |   21.016 |   694.33 |

main: n_kv_max = 18432, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.104 |   961.58 |    1.767 |    72.44 |    1.871 |   121.87 |
|   100 |    128 |    2 |    456 |    0.288 |   694.82 |    2.496 |   102.55 |    2.784 |   163.78 |
|   100 |    128 |    3 |    684 |    0.378 |   793.16 |    2.554 |   150.37 |    2.932 |   233.29 |
|   100 |    128 |    4 |    912 |    0.453 |   882.05 |    2.656 |   192.76 |    3.110 |   293.29 |
|   100 |    128 |    5 |   1140 |    0.385 |  1298.65 |    3.020 |   211.96 |    3.405 |   334.85 |
|   100 |    128 |    6 |   1368 |    0.526 |  1140.99 |    3.040 |   252.65 |    3.566 |   383.66 |
|   100 |    128 |    7 |   1596 |    0.633 |  1106.52 |    3.203 |   279.77 |    3.835 |   416.14 |
|   100 |    128 |    8 |   1824 |    0.696 |  1148.77 |    3.268 |   313.32 |    3.965 |   460.07 |
|   100 |    128 |   16 |   3648 |    1.275 |  1255.16 |    5.398 |   379.40 |    6.673 |   546.71 |
|   100 |    128 |   32 |   7296 |    2.637 |  1213.56 |    9.070 |   451.59 |   11.707 |   623.22 |
|   100 |    128 |   64 |  14592 |    6.431 |   995.24 |   17.729 |   462.08 |   24.159 |   603.99 |


LLAMA_CUBLAS=1 make -j && CUDA_VISIBLE_DEVICES=0   ./batched-bench models/codellama-7b/ggml-model-f16.gguf 18432 0 99 1 100 128 1,2,3,4,5,6,7,8,16,32,64
LLAMA_CUBLAS=1 make -j && CUDA_VISIBLE_DEVICES=0,1 ./batched-bench models/codellama-7b/ggml-model-f16.gguf 18432 0 99 1 100 128 1,2,3,4,5,6,7,8,16,32,64

### #3776

main: n_kv_max = 18432, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.047 |  2115.78 |    2.870 |    44.61 |    2.917 |    78.17 |
|   100 |    128 |    2 |    456 |    0.060 |  3340.57 |    3.113 |    82.22 |    3.173 |   143.70 |
|   100 |    128 |    3 |    684 |    0.089 |  3362.55 |    3.169 |   121.16 |    3.259 |   209.90 |
|   100 |    128 |    4 |    912 |    0.120 |  3340.32 |    3.237 |   158.18 |    3.357 |   271.70 |
|   100 |    128 |    5 |   1140 |    0.143 |  3492.60 |    3.325 |   192.48 |    3.468 |   328.70 |
|   100 |    128 |    6 |   1368 |    0.178 |  3372.40 |    3.401 |   225.83 |    3.579 |   382.26 |
|   100 |    128 |    7 |   1596 |    0.193 |  3617.95 |    3.441 |   260.42 |    3.634 |   439.17 |
|   100 |    128 |    8 |   1824 |    0.220 |  3634.74 |    3.494 |   293.07 |    3.714 |   491.09 |
|   100 |    128 |   16 |   3648 |    0.490 |  3265.68 |    4.524 |   452.66 |    5.014 |   727.52 |
|   100 |    128 |   32 |   7296 |    1.175 |  2723.82 |    6.208 |   659.77 |    7.383 |   988.21 |
|   100 |    128 |   64 |  14592 |    3.365 |  1902.21 |   14.014 |   584.54 |   17.379 |   839.64 |

main: n_kv_max = 18432, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.125 |   801.91 |    2.622 |    48.82 |    2.746 |    83.02 |
|   100 |    128 |    2 |    456 |    0.313 |   638.67 |    8.127 |    31.50 |    8.440 |    54.03 |
|   100 |    128 |    3 |    684 |    0.329 |   911.39 |    8.234 |    46.64 |    8.563 |    79.88 |
|   100 |    128 |    4 |    912 |    0.403 |   992.15 |    8.310 |    61.61 |    8.713 |   104.67 |
|   100 |    128 |    5 |   1140 |    0.439 |  1138.60 |    8.466 |    75.59 |    8.905 |   128.01 |
|   100 |    128 |    6 |   1368 |    0.586 |  1023.34 |    8.547 |    89.85 |    9.133 |   149.78 |
|   100 |    128 |    7 |   1596 |    0.737 |   950.12 |    8.744 |   102.47 |    9.481 |   168.33 |
|   100 |    128 |    8 |   1824 |    0.821 |   974.12 |    8.845 |   115.78 |    9.666 |   188.70 |
|   100 |    128 |   16 |   3648 |    1.450 |  1103.31 |   10.161 |   201.55 |   11.611 |   314.18 |
|   100 |    128 |   32 |   7296 |    3.004 |  1065.34 |   12.113 |   338.16 |   15.116 |   482.66 |
|   100 |    128 |   64 |  14592 |    7.168 |   892.86 |   20.923 |   391.54 |   28.091 |   519.46 |

### PR

main: n_kv_max = 18432, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.046 |  2168.02 |    2.873 |    44.55 |    2.919 |    78.10 |
|   100 |    128 |    2 |    456 |    0.060 |  3335.61 |    3.110 |    82.32 |    3.170 |   143.85 |
|   100 |    128 |    3 |    684 |    0.088 |  3419.15 |    3.177 |   120.85 |    3.265 |   209.48 |
|   100 |    128 |    4 |    912 |    0.119 |  3353.88 |    3.238 |   158.12 |    3.357 |   271.64 |
|   100 |    128 |    5 |   1140 |    0.145 |  3442.01 |    3.323 |   192.58 |    3.469 |   328.67 |
|   100 |    128 |    6 |   1368 |    0.179 |  3361.21 |    3.384 |   226.98 |    3.562 |   384.04 |
|   100 |    128 |    7 |   1596 |    0.194 |  3609.07 |    3.417 |   262.25 |    3.611 |   442.04 |
|   100 |    128 |    8 |   1824 |    0.222 |  3608.33 |    3.487 |   293.66 |    3.709 |   491.82 |
|   100 |    128 |   16 |   3648 |    0.492 |  3254.29 |    4.527 |   452.36 |    5.019 |   726.84 |
|   100 |    128 |   32 |   7296 |    1.174 |  2724.77 |    6.205 |   660.11 |    7.379 |   988.69 |
…