@ggerganov 2023-10-28
The reason to do it like this is because on the main device, the F16 -> F32 conversion of `dst` needs to take into account the `ldc` padding. So instead of reimplementing `to_fp32_cuda` to support a stride, I opted for this version to take advantage of `cublasGemmEx`'s dst stride support.

> As I said before, I would expect the best implementation to be converting the hidden state to FP16 on the main GPU and then distributing that.

Yes, but I have limited access to multi-GPU systems and can try only simple things like the change in this PR. Implementing F16 hidden state distribution would require more effort and resources. The long term goal is to keep all results in F16 anyway, so this will likely resolve that problem as we start to support that.