@slaren 2023-10-28
> @slaren when you implemented the use of FP16 cuBLAS GEMM, did you consider the case of >1 GPU? Because to me the code looks like for multiple GPUs FP32 is still being used on master.

I explicitly disabled the FP16 mat mul when using multiple GPUs.