@cebtenzzre 2023-10-28
This PR makes prompt processing much faster when using both my Tesla P40 and GTX 970.
| GPU      | Model     | Test | t/s master | t/s PR | Speedup |
|----------|-----------|------|------------:|--------:|---------:|
| P40 + GTX 970 | 7b q4_0   | pp512   |       156.88 |   251.06 |     1.60 |
| P40 + GTX 970 | 13b q4_k_s  | pp512   |       86.15 |   147.80 |     1.72 |