@Ph0rk0z 2023-10-29
How can this work on P40 at all since it has no tensor cores?