@cebtenzzre 2023-10-29
> How can this work on P40 at all since it has no tensor cores?

The MUL_MAT_SRC1_COL_STRIDE change applies to all GPUs. Johannes says that needs a compute capability check, but I'm not sure if he means for performance reasons or for correctness. The perplexity seems fine.