@ggerganov 2023-10-30
@cebtenzzre If you apply this patch to the current PR:

```diff
diff --git a/ggml-cuda.cu b/ggml-cuda.cu
index 9c7cf357..8e27a76d 100644
--- a/ggml-cuda.cu
+++ b/ggml-cuda.cu
@@ -6366,7 +6366,7 @@ inline void ggml_cuda_op_mul_mat_cublas(
     // ldc == nrows of the matrix that cuBLAS writes into
     const int ldc = dst->backend == GGML_BACKEND_GPU && id == g_main_device ? ne0 : row_diff;
 
-    const bool is_split = row_diff != src0->ne[1];
+    const bool is_split = false;
 
     if (compute_capability >= CC_VOLTA && (src0->type == GGML_TYPE_F16 || ggml_is_quantized(src0->type)) && ggml_is_contiguous(src0)) {
         // convert src0 and src1 to fp16, multiply as fp16, convert dst to fp32
 ggerganov ▶ Georgis-MacBook-Pro ▶ ~/development/github/llama.cpp ▶
```

**and run on single GPU (P40 or GTX970) without `GGML_CUDA_FORCE_MMQ`**, how does it compare to the PP speed **on `master` with `GGML_CUDA_FORCE_MMQ`**?