@lxrite 2023-10-30
Tested on A10 * 2. It seems that the speed of both S_PP and S_TG has been improved.
### Master 2 GPU https://github.com/ggerganov/llama.cpp/commit/6e08281e588bbba1a5d180290a94a43f167f3a1a
``` bash
CUDA_VISIBLE_DEVICES=0,1 ./batched-bench Mistral-7B-Instruct-v0.1/ggml-model-f16.gguf 14592 0 99 1 100 128 1,2,3,4,5,6,7,8,16,32,64
```
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.261 |   383.16 |    3.471 |    36.87 |    3.732 |    61.09 |
|   100 |    128 |    2 |    456 |    0.537 |   372.30 |   11.854 |    21.60 |   12.391 |    36.80 |
|   100 |    128 |    3 |    684 |    0.857 |   349.99 |   12.185 |    31.51 |   13.043 |    52.44 |
|   100 |    128 |    4 |    912 |    0.973 |   411.09 |   12.503 |    40.95 |   13.476 |    67.68 |
|   100 |    128 |    5 |   1140 |    0.992 |   504.14 |   12.654 |    50.58 |   13.645 |    83.55 |
|   100 |    128 |    6 |   1368 |    1.474 |   407.08 |   12.793 |    60.03 |   14.267 |    95.89 |
|   100 |    128 |    7 |   1596 |    1.499 |   467.04 |   13.330 |    67.21 |   14.829 |   107.63 |
|   100 |    128 |    8 |   1824 |    1.791 |   446.78 |   13.665 |    74.94 |   15.456 |   118.02 |
|   100 |    128 |   16 |   3648 |    3.404 |   470.03 |   16.306 |   125.60 |   19.710 |   185.08 |
|   100 |    128 |   32 |   7296 |    6.885 |   464.78 |   20.632 |   198.53 |   27.517 |   265.15 |
|   100 |    128 |   64 |  14592 |   14.339 |   446.34 |   37.060 |   221.05 |   51.399 |   283.90 |
### PR 2 GPU
``` bash
CUDA_VISIBLE_DEVICES=0,1 ./batched-bench Mistral-7B-Instruct-v0.1/ggml-model-f16.gguf 14592 0 99 1 100 128 1,2,3,4,5,6,7,8,16,32,64
```
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.200 |   499.45 |    3.604 |    35.52 |    3.804 |    59.94 |
|   100 |    128 |    2 |    456 |    0.495 |   404.17 |    4.172 |    61.36 |    4.667 |    97.71 |
|   100 |    128 |    3 |    684 |    0.758 |   395.83 |    4.570 |    84.02 |    5.328 |   128.38 |
|   100 |    128 |    4 |    912 |    0.800 |   500.22 |    4.759 |   107.60 |    5.558 |   164.08 |
|   100 |    128 |    5 |   1140 |    0.985 |   507.70 |    5.145 |   124.38 |    6.130 |   185.96 |
|   100 |    128 |    6 |   1368 |    1.220 |   491.65 |    5.089 |   150.93 |    6.309 |   216.83 |
|   100 |    128 |    7 |   1596 |    1.426 |   490.96 |    5.588 |   160.35 |    7.014 |   227.56 |
|   100 |    128 |    8 |   1824 |    1.471 |   543.69 |    5.952 |   172.05 |    7.423 |   245.71 |
|   100 |    128 |   16 |   3648 |    3.113 |   513.94 |    8.284 |   247.23 |   11.397 |   320.08 |
|   100 |    128 |   32 |   7296 |    6.227 |   513.91 |   13.038 |   314.15 |   19.265 |   378.72 |
|   100 |    128 |   64 |  14592 |   13.352 |   479.34 |   28.825 |   284.20 |   42.176 |   345.98 |

But it still slower than on 1 GPU.
### Master 1 GPU https://github.com/ggerganov/llama.cpp/commit/6e08281e588bbba1a5d180290a94a43f167f3a1a
``` bash
CUDA_VISIBLE_DEVICES=0 ./batched-bench Mistral-7B-Instruct-v0.1/ggml-model-f16.gguf 14592 0 99 1 100 128 1,2,3,4,5,6,7,8,16,32,64
```
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.056 |  1770.95 |    4.246 |    30.15 |    4.302 |    53.00 |
|   100 |    128 |    2 |    456 |    0.085 |  2363.90 |    4.528 |    56.54 |    4.612 |    98.87 |
|   100 |    128 |    3 |    684 |    0.123 |  2432.50 |    4.564 |    84.14 |    4.687 |   145.93 |
|   100 |    128 |    4 |    912 |    0.169 |  2365.58 |    4.646 |   110.20 |    4.815 |   189.41 |
|   100 |    128 |    5 |   1140 |    0.236 |  2121.89 |    4.716 |   135.71 |    4.952 |   230.23 |
|   100 |    128 |    6 |   1368 |    0.253 |  2375.43 |    4.809 |   159.71 |    5.061 |   270.29 |
|   100 |    128 |    7 |   1596 |    0.258 |  2715.96 |    5.048 |   177.50 |    5.306 |   300.81 |
|   100 |    128 |    8 |   1824 |    0.290 |  2756.22 |    5.268 |   194.39 |    5.558 |   328.18 |
|   100 |    128 |   16 |   3648 |    0.663 |  2414.86 |    6.575 |   311.50 |    7.237 |   504.07 |
|   100 |    128 |   32 |   7296 |    1.572 |  2035.24 |    8.468 |   483.72 |   10.040 |   726.70 |
|   100 |    128 |   64 |  14592 |    4.656 |  1374.63 |   18.453 |   443.93 |   23.109 |   631.44 |