@Ph0rk0z 2023-10-30
I did dual P40s with this PR + patch and had a regression in token speed (down to 5.5t/s). No real improvement for prompt processing or anything. That is without forcing MMV on. With it, it behaves like master.

Dual 3090 had a regression of speed from 18.93t/s down to 17.0 t/s and crashed on long context as I never took the patch off.