@esmeetu 2023-10-30
Test onto dual T4

CUDA_VISIBLE_DEVICES=0,1 ./batched-bench ./CodeLlama-7b-hf/ggml-model-f16.gguf 14592 0 99 1 100 128 1,2,3,4,5,6,7,8,16,32,64

master:

main: n_kv_max = 14592, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.386 |   259.01 |    4.130 |    30.99 |    4.516 |    50.49 |
|   100 |    128 |    2 |    456 |    0.680 |   294.17 |   18.867 |    13.57 |   19.546 |    23.33 |
|   100 |    128 |    3 |    684 |    0.909 |   330.00 |   19.402 |    19.79 |   20.311 |    33.68 |
|   100 |    128 |    4 |    912 |    1.217 |   328.81 |   19.750 |    25.92 |   20.966 |    43.50 |
|   100 |    128 |    5 |   1140 |    1.447 |   345.50 |   20.803 |    30.77 |   22.250 |    51.24 |
|   100 |    128 |    6 |   1368 |    1.839 |   326.33 |   21.261 |    36.12 |   23.100 |    59.22 |
|   100 |    128 |    7 |   1596 |    2.068 |   338.42 |   22.096 |    40.55 |   24.164 |    66.05 |
|   100 |    128 |    8 |   1824 |    2.338 |   342.11 |   22.206 |    46.11 |   24.545 |    74.31 |
|   100 |    128 |   16 |   3648 |    4.676 |   342.19 |   24.963 |    82.04 |   29.639 |   123.08 |
|   100 |    128 |   32 |   7296 |    9.761 |   327.85 |   28.503 |   143.70 |   38.264 |   190.68 |
|   100 |    128 |   64 |  14592 |   21.758 |   294.15 |   53.753 |   152.40 |   75.511 |   193.24 |

llama_print_timings:        load time =    9256.74 ms
llama_print_timings:      sample time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings: prompt eval time =  298990.17 ms / 33632 tokens (    8.89 ms per token,   112.49 tokens per second)
llama_print_timings:        eval time =    4129.63 ms /   128 runs   (   32.26 ms per token,    31.00 tokens per second)
llama_print_timings:       total time =  312069.12 ms


PR:

main: n_kv_max = 14592, is_pp_shared = 0, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   100 |    128 |    1 |    228 |    0.176 |   567.34 |    4.146 |    30.88 |    4.322 |    52.75 |
|   100 |    128 |    2 |    456 |    0.277 |   720.99 |    4.489 |    57.03 |    4.767 |    95.67 |
|   100 |    128 |    3 |    684 |    0.379 |   791.61 |    4.774 |    80.43 |    5.153 |   132.73 |
|   100 |    128 |    4 |    912 |    0.494 |   809.32 |    4.972 |   102.97 |    5.467 |   166.83 |
|   100 |    128 |    5 |   1140 |    0.592 |   845.26 |    5.495 |   116.47 |    6.087 |   187.30 |
|   100 |    128 |    6 |   1368 |    0.755 |   794.75 |    5.686 |   135.07 |    6.441 |   212.39 |
|   100 |    128 |    7 |   1596 |    0.826 |   847.94 |    6.000 |   149.33 |    6.826 |   233.82 |
|   100 |    128 |    8 |   1824 |    0.935 |   855.37 |    6.223 |   164.55 |    7.158 |   254.81 |
|   100 |    128 |   16 |   3648 |    1.986 |   805.63 |   10.519 |   194.70 |   12.505 |   291.72 |
|   100 |    128 |   32 |   7296 |    4.569 |   700.37 |   15.295 |   267.80 |   19.864 |   367.29 |
|   100 |    128 |   64 |  14592 |   11.696 |   547.21 |   38.641 |   212.00 |   50.337 |   289.89 |

llama_print_timings:        load time =    8965.64 ms
llama_print_timings:      sample time =       0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings: prompt eval time =  124935.16 ms / 33632 tokens (    3.71 ms per token,   269.20 tokens per second)
llama_print_timings:        eval time =    4145.57 ms /   128 runs   (   32.39 ms per token,    30.88 tokens per second)
llama_print_timings:       total time =  137892.88 ms