@cebtenzzre 2023-10-30
| GPU      |Model     | Test | t/s master+MMQ | t/s PR+patch | Speedup |
|----------|-----------|------|------------:|--------:|---------:|
| P40  |7b q4_0   | pp512   |    336.40    |  336.98  |   no change   |
| P40  |13b q4_k_s  | pp512   |  203.36      |  204.21  |   no change  |

FWIW, using the GTX 970 disables mmq due to lack of dp4a, and I cannot fit a 7b q4_0 on it since it has only 4GB of VRAM. But as of 2f9ec7e271220a78fe27c9e6ccbcc0dda31cda0f, there is no clear way to disable mmq in order to make a fair comparison.